rename api related salespople directives
