+++ src/app/api/data/http/http-data-api.module.ts
@@ -6,7 +6,7 @@
 import { NgModule } from '@angular/core';
 import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
 import { CustomersHttpDataApiService } from './impl/customers.http-data-api.service';
-import { SellersHttpDataApiService } from './impl/sellers.http-data-api.service';
+import { SalespeopleHttpDataApiService } from './impl/salespeople.http-data-api.service';
 import { ProductsHttpDataApiService } from './impl/products.http-data-api.service';
 import { SalesHttpDataApiService } from './impl/sales.http-data-api.service';
 import { SharedHttpDataApiService } from './impl/shared.http-data-api.service';
@@ -26,7 +26,7 @@ import { HttpDataAccessApiService } from './http-data-access-api.service';
   ],
   providers: [
     { provide: API_SERVICE_INJECTION_TOKENS.customersCrud, useClass: CustomersHttpDataApiService },
-    { provide: API_SERVICE_INJECTION_TOKENS.salespeopleCrud, useClass: SellersHttpDataApiService },
+    { provide: API_SERVICE_INJECTION_TOKENS.salespeopleCrud, useClass: SalespeopleHttpDataApiService },
     { provide: API_SERVICE_INJECTION_TOKENS.productsCrud, useClass: ProductsHttpDataApiService },
     { provide: API_SERVICE_INJECTION_TOKENS.shared, useClass: SharedHttpDataApiService },
     { provide: API_SERVICE_INJECTION_TOKENS.usersCrud, useClass: UsersHttpDataApiService },
